+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "Lua.diagnostics.globals": [
-        "LOC"
+        "LOC",
+        "_PLUGIN"
     ]
 }
